+++ src/common/constants.ts
@@ -1,1 +1,4 @@
 export const CLIENT_ID = 'node-js';
+export const SCENE_PATTERN = RegExp(
+  '^workspaces/([a-z0-9-_]{1,61})/(characters|scenes)/([a-z0-9-_]{0,61})$',
+);