+++ src/services/connection.service.ts
@@ -18,7 +18,6 @@ import { Scene } from '../entities/scene.entity';
 import { Session } from '../entities/session.entity';
 import { SessionToken } from '../entities/session_token.entity';
 import { EventFactory } from '../factories/event';
-import { TokenClientGrpcService } from './gprc/token_client_grpc.service';
 import { WorldEngineClientGrpcService } from './gprc/world_engine_client_grpc.service';
 
 interface ConnectionProps<InworldPacketT> {
@@ -57,7 +56,6 @@ export class ConnectionService<
   private intervals: NodeJS.Timeout[] = [];
   private packetQueue: QueueItem[] = [];
 
-  private tokenService = new TokenClientGrpcService();
   private engineService = new WorldEngineClientGrpcService();
 
   private onDisconnect: () => void;
@@ -96,8 +94,9 @@ export class ConnectionService<
   }
 
   async generateSessionToken() {
-    const proto = await this.tokenService.generateSessionToken(
+    const proto = await this.engineService.generateSessionToken(
       this.connectionProps.apiKey,
+      this.connectionProps.name,
     );
 
     return SessionToken.fromProto(proto);