+++ src/services/gprc/world_engine_client_grpc.service.ts
@@ -2,17 +2,20 @@ import { credentials, Metadata, ServiceError } from '@grpc/grpc-js';
 import { InworldPacket as ProtoPacket } from '@proto/packets_pb';
 import { WorldEngineClient } from '@proto/world-engine_grpc_pb';
 import {
+  AccessToken,
   CapabilitiesRequest,
   ClientRequest,
+  GenerateTokenRequest,
   LoadSceneRequest,
   LoadSceneResponse,
   UserRequest,
 } from '@proto/world-engine_pb';
 import { promisify } from 'util';
 
+import { KeySignature } from '../../auth/key_signature';
 import { Config } from '../../common/config';
-import { CLIENT_ID } from '../../common/constants';
-import { Awaitable } from '../../common/data_structures';
+import { CLIENT_ID, SCENE_PATTERN } from '../../common/constants';
+import { ApiKey, Awaitable } from '../../common/data_structures';
 import { grpcOptions } from '../../common/helpers';
 import { Logger } from '../../common/logger';
 import { SessionToken } from '../../entities/session_token.entity';
@@ -34,17 +37,40 @@ export interface SessionProps {
 
 export class WorldEngineClientGrpcService {
   private readonly config = Config.getInstance();
-  private readonly address = this.config.getEngineHost();
-  private readonly ssl = this.config.getEngineSsl();
-  private readonly credentials = this.ssl
-    ? credentials.createSsl()
-    : credentials.createInsecure();
+  private readonly address = this.config.getHost();
+  private readonly ssl = this.config.getSsl();
   private readonly grpcOptions = { ...grpcOptions };
   private readonly client = new WorldEngineClient(
-    this.address,
-    this.credentials,
-    this.grpcOptions,
+    this.config.getHost(),
+    this.ssl
+      ? credentials.createSsl()
+      : credentials.createInsecure(),
+    { ...grpcOptions },
   );
+
+  public async generateSessionToken(
+    apiKey: ApiKey,
+    name: string,
+  ): Promise<AccessToken> {
+    const metadata = new Metadata();
+    const request = new GenerateTokenRequest();
+    const resource = `workspaces/${SCENE_PATTERN.exec(name)[1]}`;
+
+    request.setKey(apiKey.key);
+    request.setResourcesList([resource]);
+    metadata.add(
+      'authorization',
+      KeySignature.getAuthorization({
+        apiKey,
+        host: this.config.getHost(),
+      }),
+    );
+
+    return promisify(this.client.generateToken.bind(this.client))(
+      request,
+      metadata,
+    );
+  }
 
   private logger = Logger.getInstance();